+++ src/hooks/useDebounce.ts
@@ -1,0 +1,20 @@
+// https://usehooks.com/useDebounce/
+import { useEffect, useState } from 'react'
+
+function useDebounce<T>(value: T, delay: number): T {
+    const [debouncedValue, setDebouncedValue] = useState<T>(value)
+
+    useEffect(() => {
+        const handler = setTimeout(() => {
+            setDebouncedValue(value)
+        }, delay)
+
+        return (): void => {
+            clearTimeout(handler)
+        }
+    }, [value, delay])
+
+    return debouncedValue
+}
+
+export { useDebounce }
